Report command failures due to exceptions
