+++ lambdas-and-strings/console-runner/src/test/java/info/javaspec/console/MainTest.java
@@ -24,14 +24,14 @@ public class MainTest {
 
     @Test
     public void runsTheCommand() throws Exception {
-      Mockito.stub(this.command.run()).toReturn(Command.Result.success());
+      Mockito.stub(this.command.run()).toReturn(Result.success());
       subject.runCommand(this.command);
       Mockito.verify(this.command, Mockito.times(1)).run();
     }
 
     @Test
     public void exitsWithTheExitCodeReturnedByTheCommand() throws Exception {
-      Command.Result failure = Command.Result.failure(42, "...you're not going to like it.");
+      Result failure = Result.failure(42, "...you're not going to like it.");
       Mockito.stub(this.command.run()).toReturn(failure);
       subject.runCommand(this.command);
       Mockito.verify(this.system, Mockito.times(1)).exit(42);
@@ -39,7 +39,7 @@ public class MainTest {
 
     @Test
     public void reportsTheResult() throws Exception {
-      Command.Result result = Mockito.mock(Command.Result.class);
+      Result result = Mockito.mock(Result.class);
       Mockito.stub(this.command.run()).toReturn(result);
       subject.runCommand(this.command);
       Mockito.verify(result).reportTo(Mockito.same(this.reporter));
